fix(module: menu): submenu montion & style (#3024)
* fix(module: menu): submenu open montion & style

* fix test

+++ components/core/Component/Overlay/Overlay.razor.cs
@@ -459,7 +459,7 @@ namespace AntDesign.Internal
             }
         }
 
-        private int ChangeOverlayLeftToRight(int left, HtmlElement overlay, HtmlElement container)
+        private decimal ChangeOverlayLeftToRight(int left, HtmlElement overlay, HtmlElement container)
         {
             return container.ClientWidth - left - overlay.OffsetWidth;
         }
